Adds plugin not found component and unlisted plugins constant.
