+++ client/my-sites/plugins/constants.js
@@ -77,3 +77,5 @@ export const ECOMMERCE_BUNDLED_PLUGINS = [
 	'tiktok-for-woocommerce',
 	'mailpoet-business',
 ];
+
+export const UNLISTED_PLUGINS = [ 'automated-db-schenker-shipping ' ];
