+++ client/my-sites/plugins/plugin-not-found/index.jsx
@@ -1,0 +1,26 @@
+import { useTranslate } from 'i18n-calypso';
+import { useSelector } from 'react-redux';
+import EmptyContent from 'calypso/components/empty-content';
+import MainComponent from 'calypso/components/main';
+import { getSelectedSite } from 'calypso/state/ui/selectors';
+
+const PluginDoesNotExistView = () => {
+	const translate = useTranslate();
+	const selectedSite = useSelector( getSelectedSite );
+	const actionUrl = '/plugins' + ( selectedSite ? '/' + selectedSite.slug : '' );
+	const action = translate( 'Browse all plugins' );
+
+	return (
+		<MainComponent wideLayout>
+			<EmptyContent
+				title={ translate( "Oops! We can't find this plugin!" ) }
+				line={ translate( "The plugin you are looking for doesn't exist." ) }
+				actionURL={ actionUrl }
+				action={ action }
+				illustration="/calypso/images/illustrations/illustration-404.svg"
+			/>
+		</MainComponent>
+	);
+};
+
+export default PluginDoesNotExistView;
